requirements.txt
fastapi
uvicorn
psycopg2-binary
Faker
jinja2